fix(common): 调整源码集结构；提取 time 相关内容为独立模块
