+++ settings.gradle.kts
@@ -72,6 +72,7 @@ include(":simbot-commons:simbot-common-core")
 include(":simbot-commons:simbot-common-ktor-inputfile")
 include(":simbot-commons:simbot-common-suspend-runner")
 include(":simbot-commons:simbot-common-stage-loop")
+include(":simbot-commons:simbot-common-time")
 include(":simbot-api")
 include(":simbot-test")
 include(":simbot-cores:simbot-core")
