--- simbot-commons/simbot-common-core/src/jsMain/kotlin/love/forte/simbot/common/coroutines/Dispatchers.js.kt
@@ -1,35 +1,0 @@
-/*
- *     Copyright (c) 2024. ForteScarlet.
- *
- *     Project    https://github.com/simple-robot/simpler-robot
- *     Email      ForteScarlet@163.com
- *
- *     This file is part of the Simple Robot Library (Alias: simple-robot, simbot, etc.).
- *
- *     This program is free software: you can redistribute it and/or modify
- *     it under the terms of the GNU Lesser General Public License as published by
- *     the Free Software Foundation, either version 3 of the License, or
- *     (at your option) any later version.
- *
- *     This program is distributed in the hope that it will be useful,
- *     but WITHOUT ANY WARRANTY; without even the implied warranty of
- *     MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- *     Lesser GNU General Public License for more details.
- *
- *     You should have received a copy of the Lesser GNU General Public License
- *     along with this program.  If not, see <https://www.gnu.org/licenses/>.
- *
- */
-
-package love.forte.simbot.common.coroutines
-
-import kotlinx.coroutines.CoroutineDispatcher
-import kotlinx.coroutines.Dispatchers
-
-/**
- * 得到 [Dispatchers.Default]。
- *
- * @see Dispatchers.Default
- */
-public actual inline val Dispatchers.IOOrDefault: CoroutineDispatcher
-    get() = Default
