--- simbot-commons/simbot-common-core/src/jvmMain/kotlin/love/forte/simbot/common/time/TimeUnit.jvm.kt
@@ -1,31 +1,0 @@
-/*
- *     Copyright (c) 2023-2024. ForteScarlet.
- *
- *     Project    https://github.com/simple-robot/simpler-robot
- *     Email      ForteScarlet@163.com
- *
- *     This file is part of the Simple Robot Library.
- *
- *     This program is free software: you can redistribute it and/or modify
- *     it under the terms of the GNU Lesser General Public License as published by
- *     the Free Software Foundation, either version 3 of the License, or
- *     (at your option) any later version.
- *
- *     This program is distributed in the hope that it will be useful,
- *     but WITHOUT ANY WARRANTY; without even the implied warranty of
- *     MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- *     Lesser GNU General Public License for more details.
- *
- *     You should have received a copy of the Lesser GNU General Public License
- *     along with this program.  If not, see <https://www.gnu.org/licenses/>.
- *
- */
-
-package love.forte.simbot.common.time
-
-/**
- * 时间单位，用于时间转化。
- *
- * @see java.util.concurrent.TimeUnit
- */
-public actual typealias TimeUnit = java.util.concurrent.TimeUnit
