--- simbot-commons/simbot-common-core/src/wasmJsMain/kotlin/love/forte/simbot/common/time/Date.kt
@@ -1,38 +1,0 @@
-/*
- *     Copyright (c) 2024-2025. ForteScarlet.
- *
- *     Project    https://github.com/simple-robot/simpler-robot
- *     Email      ForteScarlet@163.com
- *
- *     This file is part of the Simple Robot Library (Alias: simple-robot, simbot, etc.).
- *
- *     This program is free software: you can redistribute it and/or modify
- *     it under the terms of the GNU Lesser General Public License as published by
- *     the Free Software Foundation, either version 3 of the License, or
- *     (at your option) any later version.
- *
- *     This program is distributed in the hope that it will be useful,
- *     but WITHOUT ANY WARRANTY; without even the implied warranty of
- *     MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- *     Lesser GNU General Public License for more details.
- *
- *     You should have received a copy of the Lesser GNU General Public License
- *     along with this program.  If not, see <https://www.gnu.org/licenses/>.
- *
- */
-
-package love.forte.simbot.common.time
-
-/**
- * Represents a date object capable of providing the current epoch time in milliseconds.
- * This class is designed to work with JavaScript environments and provides a bridge
- * to access the `getTime` method which retrieves the number of milliseconds since
- * January 1, 1970, 00:00:00 UTC.
- *
- * Due to its external nature, instances of this class should be obtained through
- * platform-specific methods rather than instantiated directly.
- */
-@OptIn(ExperimentalWasmJsInterop::class)
-internal external class Date : JsAny {
-    fun getTime(): Double
-}
