+++ simbot-commons/simbot-common-time/src/jvmMain/java/module-info.java
@@ -1,0 +1,5 @@
+module simbot.common.time {
+    requires kotlin.stdlib;
+
+    exports love.forte.simbot.common.time;
+}
